requirements.txt
Django==4.1 
psycopg2-binary==2.9.3
djangorestframework==3.14.0
markdown==3.4.1
django-filter==22.1
